Add patch for preventDefault to library.
