+++ src/library.js
@@ -9,8 +9,11 @@ import fastaParser from "./helpers/fasta";
 import computeLabelWidth from "./helpers/computeLabelWidth";
 import ScrollBroadcaster from "./helpers/ScrollBroadcaster";
 import * as colors from "./helpers/colors";
+import PreventDefaultPatch from "./prevent_default_patch";
 
 require("./app/styles.scss");
+
+PreventDefaultPatch(document);
 
 export default Alignment;
 export {
